Add preliminary support for Seccomp
Seccomp makes bugs in jq harder to exploit by imposing limits on the
behaviour of the binary. On transgressions the Linux kernel will kill
it, and print "Bad system call" (or similar). (The kernel's audit log
will contain more details.)

This patch comes with a prelimiary ruleset. That is, I might have missed
some legal syscalls.